backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
pydantic[email]
python-multipart
python-jose[cryptography]
passlib[argon2]
python-dotenv
feedparser
requests
python-dateutil
jinja2
httpx
python-docx
semgrep
groq
scikit-learn
numpy
boto3
google-cloud-iam
google-cloud-resource-manager